Main: keyboard HID usage ID is 16-bit.
svn:sync-xref-src-repo-rev: r133940

+++ src/VBox/Main/src-client/KeyboardImpl.cpp
@@ -1,4 +1,4 @@
-/* $Id: KeyboardImpl.cpp 79738 2019-07-12 13:59:54Z michal.necasek@oracle.com $ */
+/* $Id: KeyboardImpl.cpp 81246 2019-10-14 10:00:06Z michal.necasek@oracle.com $ */
 /** @file
  * VirtualBox COM class implementation
  */
@@ -255,7 +255,7 @@ HRESULT Keyboard::putUsageCode(LONG aUsageCode, LONG aUsagePage, BOOL fKeyReleas
 
     int vrc = VINF_SUCCESS;
     uint32_t u32Usage;
-    u32Usage = (uint8_t)aUsageCode | ((uint32_t)(uint8_t)aUsagePage << 16) | (fKeyRelease ? 0x80000000 : 0);
+    u32Usage = (uint16_t)aUsageCode | ((uint32_t)(uint8_t)aUsagePage << 16) | (fKeyRelease ? 0x80000000 : 0);
     vrc = pUpPort->pfnPutEventHid(pUpPort, u32Usage);
     if (RT_FAILURE(vrc))
         return setErrorBoth(VBOX_E_IPRT_ERROR, vrc,
